backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
Pillow==10.2.0
python-dotenv==1.0.0
marshmallow==3.20.2
mysqlclient==2.2.3
PyMySQL==1.1.0
gunicorn==21.2.0
nltk==3.8.1
supabase==2.1.4
